code/requirements.txt
torch>=2.1
numpy>=1.24
pillow>=10.0
tqdm>=4.66

